move graticule break calculation
